(fix): refactor main menu
